+++ Stripe/StripeiOSTests/STPPaymentMethodFunctionalTest.swift
@@ -155,4 +155,17 @@ class STPPaymentMethodFunctionalTest: XCTestCase {
 
         waitForExpectations(timeout: 5, handler: nil)
     }
+    
+    func testCreateMobilePayPaymentMethod() {
+        let client = STPAPIClient(publishableKey: STPTestingFRPublishableKey)
+        let params = STPPaymentMethodParams(mobilePay: STPPaymentMethodMobilePayParams(), billingDetails: nil, metadata: nil)
+        let expectation = self.expectation(description: "Payment Method create")
+        client.createPaymentMethod(with: params) { paymentMethod, error in
+            XCTAssertNil(error)
+            XCTAssertNotNil(paymentMethod)
+            XCTAssertEqual(paymentMethod?.type, .mobilePay)
+            expectation.fulfill()
+        }
+        waitForExpectations(timeout: 5, handler: nil)
+    }
 }
